+++ Doom3/Source/Core/DObject/DObjectManager.cpp
@@ -7,24 +7,24 @@
 #include <chrono>
 #include <UI/PrintText.h>
 
-unsigned long long doom::DObjectManager::GenerateNewDObejctID()
+UINT64 doom::DObjectManager::GenerateNewDObejctID()
 {
-    const unsigned long long newDObjectID = mDObjectCounter++;
+    const UINT64 newDObjectID = mDObjectCounter++;
    
     assert(newDObjectID != 0xFFFFFFFFFFFFFFFF);
 
     return newDObjectID;
 }
 
-std::unordered_map<doom::DObject*, unsigned long long>::iterator  doom::DObjectManager::SetDObjectID(DObject* const dObject, const unsigned long long dObjectID)
+std::unordered_map<doom::DObject*, UINT64>::iterator  doom::DObjectManager::SetDObjectID(DObject* const dObject, const UINT64 dObjectID)
 {
     assert(dObject != nullptr);
 
-    std::unordered_map<doom::DObject*, unsigned long long>::iterator iter = mDObjectsList.end();
+    std::unordered_map<doom::DObject*, UINT64>::iterator iter = mDObjectsList.end();
 
     if (dObject != nullptr)
     {
-        std::pair< std::unordered_map<doom::DObject*, unsigned long long>::iterator, bool> result = mDObjectsList.insert_or_assign(dObject, dObjectID);
+        std::pair< std::unordered_map<doom::DObject*, UINT64>::iterator, bool> result = mDObjectsList.insert_or_assign(dObject, dObjectID);
         iter = result.first;
 
 		dObject->mDObjectID = dObjectID;
@@ -41,7 +41,7 @@ bool doom::DObjectManager::AddNewDObject(DObject* const dObject)
 
     if (dObject != nullptr)
 	{
-        const unsigned long long newDObjectID = DObjectManager::GenerateNewDObejctID();
+        const UINT64 newDObjectID = DObjectManager::GenerateNewDObejctID();
         SetDObjectID(dObject, newDObjectID);
 
         isSuccess = true;
@@ -56,7 +56,7 @@ bool doom::DObjectManager::ReplaceDObjectFromDObjectList(DObject&& originalDObje
     assert(originalDObject.GetDObjectID() != INVALID_DOBJECT_ID);
     assert(&originalDObject != newDObject);
 
-    const unsigned long long originalDObjectID = originalDObject.GetDObjectID();
+    const UINT64 originalDObjectID = originalDObject.GetDObjectID();
 
     SetDObjectID(newDObject, originalDObjectID);
 
@@ -67,9 +67,9 @@ bool doom::DObjectManager::ReplaceDObjectFromDObjectList(DObject&& originalDObje
 
 bool doom::DObjectManager::RemoveDObject(DObject* const dObject)
 {
-    const unsigned long long dObjectID = dObject->GetDObjectID();
+    const UINT64 dObjectID = dObject->GetDObjectID();
     
-    std::unordered_map<doom::DObject*, unsigned long long>::iterator targetIter = SetDObjectID(dObject, INVALID_DOBJECT_ID);
+    std::unordered_map<doom::DObject*, UINT64>::iterator targetIter = SetDObjectID(dObject, INVALID_DOBJECT_ID);
     if (targetIter != mDObjectsList.end())
     {
         targetIter = mDObjectsList.erase(targetIter); //erasing iterator is much faster than not erasing
@@ -81,13 +81,13 @@ bool doom::DObjectManager::RemoveDObject(DObject* const dObject)
 
 void doom::DObjectManager::DestroyAllDObjects(const bool force)
 {
-    std::unordered_map<doom::DObject*, unsigned long long>::iterator erasedIter = mDObjectsList.begin();
+    std::unordered_map<doom::DObject*, UINT64>::iterator erasedIter = mDObjectsList.begin();
 
     while (erasedIter != mDObjectsList.end())
     {
 		if (erasedIter->second != INVALID_DOBJECT_ID && ( force || erasedIter->first->GetDObjectFlag(eDObjectFlag::NewAllocated) == true ) )
 		{
-            //std::unordered_map<doom::DObject*, unsigned long long>::iterator deletedIter = erasedIter++;
+            //std::unordered_map<doom::DObject*, UINT64>::iterator deletedIter = erasedIter++;
 			//delete deletedIter->first;
 
 			delete erasedIter->first;
@@ -106,7 +106,7 @@ bool doom::DObjectManager::IsDObjectValid(const DObject* const dObject)
 
 	if (dObject != nullptr)
 	{
-		std::unordered_map<DObject*, unsigned long long>::const_iterator iter = mDObjectsList.find(const_cast<DObject *>(dObject));
+		std::unordered_map<DObject*, UINT64>::const_iterator iter = mDObjectsList.find(const_cast<DObject *>(dObject));
 		if (iter != mDObjectsList.end() && iter->second != INVALID_DOBJECT_ID)
 		{
 			isValid = true;
@@ -122,7 +122,7 @@ bool doom::DObjectManager::IsEmpty()
     return mDObjectsList.empty();
 }
 
-size_t doom::DObjectManager::GetDObjectCount()
+SIZE_T doom::DObjectManager::GetDObjectCount()
 {
     return mDObjectsList.size();
 }
