+++ Doom3/Source/Core/Physics/Collider/RenderCollider.h
@@ -12,7 +12,7 @@ namespace doom
 		{
 		private:
 
-			unsigned long long mLastRenderFrameTickCount{};
+			UINT64 mLastRenderFrameTickCount{};
 
 		protected:
